refactor(types): deprecate Issue file
It was really confusing

Use only src/types

Create a types file for issue-tracker (private)
